time_intervals function of Spike Cataloger changed
time_intervals function now takes the arguments x and z instead of x and
y to eliminate confusion with the data, since the data is stored in the
variable y.  Attempted to add a code that would allow time_intervals to
find the peak of a spike, but currently it always returns nan.