Spike_Cataloger returns a data frame for all scopes/channels in any shot now
The program loops through all scopes and channels in a shot now.
Updated functions where necessary to accommodate this.  Notably,
list_to_frame works with a list of lists now.  This list of lists is
stored as final_results.  The program appears to work, but it currently
says that there is only one spike per channel in any given scope and
that seems wrong.